requirements.txt
langchain
openai
python-dotenv
langchain_community
pypdf
yt_dlp
pydub
ffprobe 
ffmpeg 
transformers 
datasets 
torch 
faiss-cpu 
wget
matplotlib 
scikit-learn
chromadb
lark
panel
langchain-openai 
docarray
jupyter_bokeh
langchain_chroma
streamlit